+++ source/ModelGpu/_Impl/Cuda/CudaMap.cuh
@@ -7,15 +7,15 @@ __host__ __device__ void mapPosCorrection(int2 &pos, int2 const &size)
 	pos = { ((pos.x % size.x) + size.x) % size.x, ((pos.y % size.y) + size.y) % size.y };
 }
 
-__host__ __device__ void mapPosCorrection(double2 &pos, int2 const &size)
+__host__ __device__ void mapPosCorrection(float2 &pos, int2 const &size)
 {
 	int2 intPart{ (int)pos.x, (int)pos.y };
-	double2 fracPart = { pos.x - intPart.x, pos.y - intPart.y };
+	float2 fracPart = { pos.x - intPart.x, pos.y - intPart.y };
 	mapPosCorrection(intPart, size);
-	pos = { static_cast<double>(intPart.x) + fracPart.x, static_cast<double>(intPart.y) + fracPart.y };
+	pos = { static_cast<float>(intPart.x) + fracPart.x, static_cast<float>(intPart.y) + fracPart.y };
 }
 
-__device__ void mapDisplacementCorrection_Kernel(double2 &disp, int2 const &size)
+__device__ void mapDisplacementCorrection_Kernel(float2 &disp, int2 const &size)
 {
 }
 
@@ -41,9 +41,9 @@ __host__ __device__ void inline setCellToMap(int2 posInt, CellCuda * __restrict_
 	map[mapEntry] = cell;
 }
 
-__device__ inline double mapDistanceSquared_Kernel(double2 const &p, double2 const &q, int2 const &size)
+__device__ inline float mapDistanceSquared_Kernel(float2 const &p, float2 const &q, int2 const &size)
 {
-	double2 d = { p.x - q.x, p.y - q.y };
+	float2 d = { p.x - q.x, p.y - q.y };
 	mapDisplacementCorrection_Kernel(d, size);
 	return d.x*d.x + d.y*d.y;
 }
